ageRecognition/requirements/common.txt
Django==1.6.5
Fabric==1.9.1
ImageHash==0.3
MySQL-python==1.2.5
PIL==1.1.7
PyMySQL==0.6.2
South==1.0
Theano==0.6.0
Unidecode==0.04.16
argparse==1.2.1
django-allauth==0.18.0
django-analytical==0.19.0
django-appconf==0.6
django-compressor==1.4
django-config-gen==1.0.8
django-configurations==0.8
django-debug-toolbar==1.2.1
django-extensions==1.3.10
django-extra-views==0.6.5
django-facebook==6.0.1
django-guardian==1.2.4
django-haystack==2.2.0
django-js-reverse==0.3.0
django-mptt==0.6.1
django-nose==1.2
django-storages==1.1.8
django-tinymce==1.5.2
djangorestframework==2.3.14
ecdsa==0.11
neurolab==0.3.4
nose==1.3.3
numpy==1.9.0
oauthlib==0.6.3
paramiko==1.14.0
pycrypto==2.6.1
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
scipy==0.14.0
six==1.5.2
sorl-thumbnail==11.12
sqlparse==0.1.11
virtualenv==1.11.6
wsgiref==0.1.2